pyproject.toml
[project]
name = "dtecflex-rss-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.13.3",
    "argon2-cffi>=23.1.0",
    "bcrypt>=4.2.0",
    "bs4>=0.0.2",
    "emoji>=2.14.0",
    "openai>=1.51.2",
    "pandas>=2.2.3",
    "playwright>=1.48.0",
    "pydantic>=2.9.2",
    "pyjwt>=2.9.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.35",
    "streamlit-authenticator>=0.4.1",
    "streamlit-cookies-controller>=0.0.4",
    "streamlit-modal>=0.1.2",
    "streamlit>=1.39.0",
    "tiktoken>=0.8.0",
]
